backend/requirements.txt
Django==4.0.3
djangorestframework==3.13.1
psycopg2-binary==2.9.3
django-filter==21.1
django-cors-headers==3.11.0
Pillow==9.0.1
gunicorn==20.1.0
django-geojson==3.2.0
pyyaml==6.0
python-dotenv==0.19.2
twilio==7.7.0
djangorestframework-simplejwt==5.0.0
